requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
healthcheck==1.3.1
psycopg2
uwsgi>=2.0.13.1
pyyaml
